auto download shcedule update

+++ RegulusDataSetter/main.py
@@ -58,24 +58,36 @@ def download_schedule():
     url = read_config('url')
     #url = "https://dphajek-windows.azurewebsites.net/Api/Regulus/GetSchedule"
     next_date = (datetime.date.today() + datetime.timedelta(days=1)).strftime("%Y-%m-%d")
-    #next_date = (datetime.date.today()).strftime("%Y-%m-%d")
+    actual_date = datetime.date.today().strftime("%Y-%m-%d")
+    date = next_date
 
-    file_name = next_date + ".txt"
+    current_time = datetime.datetime.now().time()
+    if current_time > datetime.time(19, 00) and schedule_file_exists(next_date + ".txt"):
+        print_log("Schedule for " + next_date, "download_schedule")
+        return
+    
+    if current_time < datetime.time(18, 59) and schedule_file_exists(actual_date + ".txt"):
+        print_log("Schedule for " + actual_date, "download_schedule")
+        return
+    else:
+        date = actual_date
+
+    file_name = date + ".txt"
     if os.path.exists(file_name):
         print_log(f"File {file_name} already exists", "download_schedule")
         return
 
     while True:
-        print_log("Downloading schedule", "download_schedule")
+        print_log("Downloading schedule " + date, "download_schedule")
         try:
             params = {
-                'date': next_date,
+                'date': date,
                 'guid': guid
             }
             response = requests.get(url, params=params)
             response.raise_for_status()
             json_data = response.json()
-            file_name = next_date + ".txt"
+            file_name = date + ".txt"
             with open(file_name, 'w') as file:
                 file.write(str(json_data))
             print_log(f"JSON data downloaded and saved to {file_name}", "download_schedule")
@@ -83,6 +95,11 @@ def download_schedule():
         except (requests.RequestException, IOError) as e:
             print_log(f"Failed to download schedule {e}", "download_schedule")
             time.sleep(300)  # Wait for 5 minutes before next attempt
+
+def schedule_file_exists(file_name):
+    if os.path.exists(file_name):
+        return True
+    return False
 
 def read_config(value_name):
     try:
@@ -108,9 +125,10 @@ if __name__ == '__main__':
     scheduler = BlockingScheduler()
     scheduler.add_job(read_actual_time_actions, 'cron', minute='0,15,30,45')
     scheduler.add_job(download_schedule, 'cron', hour=19, minute=0)
+    # Zapnu hned aby se kdyžtak stáhnul aktuiální rozvrh
     current_time = datetime.datetime.now().time()
-    if current_time > datetime.time(18, 59):
-        scheduler.add_job(download_schedule, 'date', run_date=datetime.datetime.now() + datetime.timedelta(seconds=5))
+    scheduler.add_job(download_schedule, 'date', run_date=datetime.datetime.now() + datetime.timedelta(seconds=5))
+
     scheduler.start()
     scheduler.start()
